branch using spring-boot-starter-data-rest

+++ src/main/java/com/kpaw/sakilaspringbootrest/repository/movie/ActorRepository.java
@@ -6,7 +6,7 @@ import org.springframework.data.domain.Pageable;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.data.jpa.repository.Query;
 
-public interface ActorRepository extends JpaRepository<Actor, Short> {
+public interface ActorRepository extends JpaRepository<Actor, Integer> {
 
     @Query(value = "SELECT * FROM actor WHERE actor_id IN (SELECT actor_id FROM film_actor WHERE film_id = ?1)", nativeQuery = true)
     Page<Actor> findActorsByFilmId(int id, Pageable pageable);
