+++ src/main/java/com/kpaw/sakilaspringbootrest/repository/rental/CustomerRepository.java
@@ -5,7 +5,7 @@ import org.springframework.data.domain.Page;
 import org.springframework.data.domain.Pageable;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface CustomerRepository extends JpaRepository<Customer, Short> {
+public interface CustomerRepository extends JpaRepository<Customer, Integer> {
 
    Page<Customer> findCustomerByFirstNameContainsOrLastNameContainsAllIgnoreCase(String firstName, String lastName, Pageable pageable);
 }
