+++ src/main/java/com/kpaw/sakilaspringbootrest/repository/rental/StoreRepository.java
@@ -3,5 +3,5 @@ package com.kpaw.sakilaspringbootrest.repository.rental;
 import com.kpaw.sakilaspringbootrest.domain.location.Store;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface StoreRepository extends JpaRepository<Store, Byte> {
+public interface StoreRepository extends JpaRepository<Store, Integer> {
 }
